requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
ta>=0.7.0

# Binance integration
python-binance>=1.0.16
websockets>=10.0
pycryptodome>=3.10.0
aiohttp>=3.8.0

# Optional dependencies (Windows only)
# MetaTrader5>=5.0.33 ; sys_platform == 'win32'
# pymt5adapter>=0.4.4 ; sys_platform == 'win32'

# UI dependencies (may require additional setup)
# PySide2>=5.15.0 
streamlit>=1.24.0 